Touchscreen Rotate Tool v1
First Implementation of the rotation tool compatible with touchscreen devices. See the example scene within the FRONT END / AR / ARRotate Folder.

Note: The implementation uses the 'touch' class in unity - and is not directly compatible with mouse input, so will only work when you export to a mobile device.

I have duplicated a bunch of scripts during the process of figuring everything out because I didn't want to change any of Luke's work.