Base Movement
Smooth interpolation for robot position. Call the TranslateBase method for the selected robot  class instance (which can be accessed from selectionManager singleton). Pass in the target position as a Vector3. Speed and snappiness of movement can be controlled through variables maxSpeed and floatTime at top of robot class. Test scene for reference present (TranslatorTest), but make sure to comment out line27 to avoid an error from lack of selection manager in scene.